repo/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
alembic==1.16.5
pydantic-settings==2.11.0
argon2-cffi==25.1.0
cryptography==46.0.1
python-multipart==0.0.20
openpyxl==3.1.5
